Fix non linux build
