+++ client/internal/wgproxy/factory.go
@@ -1,32 +1,8 @@
 package wgproxy
 
-import (
-	"runtime"
-
-	log "github.com/sirupsen/logrus"
-)
-
 type Factory struct {
 	wgPort    int
 	ebpfProxy Proxy
-}
-
-func NewFactory(wgPort int) *Factory {
-	f := &Factory{wgPort: wgPort}
-
-	if runtime.GOOS != "linux" {
-		return f
-	}
-
-	ebpfProxy := NewWGEBPFProxy(wgPort)
-	err := ebpfProxy.Listen()
-	if err != nil {
-		log.Errorf("failed to initialize ebpf proxy: %s", err)
-		return f
-	}
-
-	f.ebpfProxy = ebpfProxy
-	return f
 }
 
 func (w *Factory) GetProxy() Proxy {
